+++ site/darklua-wasm/tests/web.rs
@@ -10,5 +10,8 @@ wasm_bindgen_test_configure!(run_in_browser);
 
 #[wasm_bindgen_test]
 fn process_empty_code() {
-    assert_eq!(darklua_wasm::process_code("", JsValue::undefined()).unwrap(), "");
+    assert_eq!(
+        darklua_wasm::process_code("", JsValue::undefined()).unwrap(),
+        ""
+    );
 }